size

+++ hello_world/6-size.c
@@ -10,12 +10,12 @@ int main(void)
 char charType;
 int intType;
 long int longintType;
-long double longlongType;
+double longlongType;
 float floatType;
- printf("Size of char: %zu byte\n", sizeof(charType));
+printf("Size of char: %zu byte\n", sizeof(charType));
 printf("Size of int: %zu bytes\n", sizeof(intType));
- printf("Size of long int: %zu bytes\n", sizeof(longintType));
- printf("Size of long long int: %zu bytes\n", sizeof(longlongType));
+printf("Size of long int: %zu bytes\n", sizeof(longintType));
+printf("Size of long long int: %zu bytes\n", sizeof(longlongType));
 printf("Size of float: %zu bytes\n", sizeof(floatType));
 return (0);
 }
